+++ server.h
@@ -1,16 +1,19 @@
 #ifndef SERVER_H
-#define	SERVER_H
+# define	SERVER_H
 
-#include "shell.h"
-#include <sys/epoll.h>
-#include <sys/socket.h>
-#include <errno.h>
-#include <netinet/in.h>
-#include <sys/select.h>
-#include <unistd.h>
-#include <stdbool.h>
+# include "shell.h"
+# include <sys/epoll.h>
+# include <sys/socket.h>
+# include <errno.h>
+# include <netinet/in.h>
+# include <sys/select.h>
+# include <unistd.h>
+# include <stdbool.h>
 
 # define PORT_FORTYTWO 4242
+# define PASSWORD "4242"
+# define SOCK_MAX 32
+# define MAX_CLIENT 3
 
 typedef struct s_server
 {
@@ -24,6 +27,19 @@ typedef struct s_server
 	int		current_client;
 }			t_server;
 
+typedef struct s_clients
+{
+	int num; // currently connected clients
+	int list[MAX_CLIENT]; // list of clients
+	int auth[MAX_CLIENT]; // list of authorized clients
+	int new; // new connected index
+	int free; // last disconnected index
+}	t_clients;
+
 void	server();
+int 	pre_shell(int fd, int epollfd, t_clients *clients);
+void	shell(int fd, int epollfd, t_clients *clients);
+void	process_input(char *str);
+void	remove_client(int fd, int epollfd, t_clients *clients);
 
 #endif
